Create README.md
Voice conversion using deep learning using the approach of STFT which capture the sound wave in an image form. The aim of the research is to make the input sound change to the style sound. Unfortunately, it works only on the parallel sound input which is the same sentence for input and style but different tone.